Handle holes with no containing polygon.
If a hole's enclosing ring cannot be found (assuming exterior rings are
clockwise and holes are counterclockwise as per the Shapefile specification),
then represent the hole as an exterior polygon.

It is much more likely that this will occur if you attempt to parse a shapefile
that is not in EPSG:4326 (spherical coordinates). The correct fix there is to
reproject the shapefile using ogr2ogr -t_srs (e.g., mbostock/topojson#54).
However, this change at least avoids a spurious error with invalid input.