Convert coordinates to WGS 84 base on the .prj file
This uses proj4js to read the "well known" text of
the coordinate reference system. Then as we read in
coordinates from the `.shp` file we map each to WGS 84
using the same proj4js library.

https://tools.ietf.org/html/rfc7946#section-4 suggests
that's the only allowed coordinate reference system in
GeoJSON.

Unfortunately this is a bit of a hack, it breaks with
streaming nature of the program. It also affects only
the command line `shp2json`. Accommodations for library
use are lacking. There should be a command line switch
for deciding to use the `.prj` transform  this or not.
Arguably the mapping should not be conflated with
reading, but it was the most convenient way for me to
make it work. Tests are absent.